ENGINE-1620: Change BLOB to MEDIUMBLOB / LONGBLOB for MySQL column type (also changed other RDBMS)

+++ bpm/bonita-core/bonita-process-instance/bonita-process-instance-model-impl/src/main/resources/org/bonitasoft/engine/core/process/instance/model/impl/db/mysql-createTables.sql
@@ -55,7 +55,7 @@ CREATE TABLE flownode_instance (
   lastUpdateDate BIGINT,
   expectedEndDate BIGINT,
   claimedDate BIGINT,
-  priority BIGINT,
+  priority TINYINT,
   gatewayType VARCHAR(50),
   hitBys VARCHAR(255),
   stateCategory VARCHAR(50) NOT NULL,
